+++ src/main/java/com/github/kd_gaming1/packcore/config/PackCoreConfig.java
@@ -83,4 +83,8 @@ public class PackCoreConfig extends MidnightConfig {
     @Hidden
     @Entry(category = META)
     public static boolean isFirstStartup = true;
+
+    @Hidden
+    @Entry(category = META)
+    public static boolean modernUIDefaultsEnforced = false;
 }